Fix multiselect widget signal handling.
Apparently list isn't allowed type of argument.
Documentation says:
"It is possible to pass any Python object as a signal argument by specifying
PyQt_PyObject as the type of the argument in the signature."
"While this would normally be used for passing objects like lists and
dictionaries as signal arguments, it can be used for any Python type."

So do it.