Qubes VM Manager drop down menu strings consistency fixes
Currently:
* Create AppVM
* Remove AppVM
* Clone VM
* Start/Resume VM
* [...] VM

The first two are inconsistent. @bnvk and I agreed, that those should be changed AppVM -> VM for consistency.

And I add, if anything, it would have to be "Create TemplateBased-VM". Because currently, if you click "Create AppVM", you are asked in the next wizard if you wanted to create an AppVM, NetVM or ProxyVM. So the term AppVM is overloaded.

This commit fixes this.