backups: Explain why passwords are always required
I had someone tell me IRL that they thought the fact that it still asked
for a password when "Encrypt backup" was unchecked was a bug. This is
not the case, so explain why.

I think this wording is more user-friendly than "also used for an HMAC"
or such.

Fixes https://github.com/QubesOS/qubes-issues/issues/2365