Minor textual update for Global Settings
In lieu of bigger overhaul/processing policy files in
VM Settings, a bit of explanation for updateVM in Global
Settings and a change of name from UpdateVM to Dom0 Update VM
there. This is quick fix solution, pending a better
one when (hopefully) we get some better API for
interacting with policies.

references QubesOS/qubes-issues#3412