global_settings: Use MiB, not M as the unit
Prior to this commit, clicking on the "OK" button in the global settings
window would cause the dom0 memory boost and minimum VM memory fields to
be saved incorrectly, due to the fact that the code calculates the
values to be saved using the "MiB" as the units (i.e., factors of
1024), whereas the code would save the corresponding string into the
qmemman configuration file with the "M" suffix (indicating factors of
1000 instead of 1024).

Due to this bug, the aforementioned fields' values would be scaled down
every time the user clicked on the "OK" button in the global settings
window. For example, if the user entered 400 and hit 'OK', opening the
global settings window again would show the value 381.