Fix KernelVersion comparisons
In the __lt__ function for the class KernelVersions, if
self.groups != other.groups, but self.groups == other.groups[0:n] or
self.groups[0:n] == other.groups for some n, then at some point, one of
the two pieces to be compared will be None, which resulted in an
Exception when calling isdigit.

Hence check whether one of the pieces to be compared is None and handle
this as a special case.