Avoid spurious warnings when setting NetVM
Qube Manager currently shows a security warning when changing the NetVM of TemplateVMs.
This warning is generally useful, but not when disabling networking;
for example changing the NetVM from from "default" to "none".

In this chnage I suppress the warning when the NetVM for a TemplateVM is changed to "none".