Fix display of kernels for Qube Settings
After setting a vm.kernel to None, the property
vm.kernel outputted an empty string, not None object,
which caused some funny problems with display of this
property for Qube Settings (namely, we got a ' (current)' instead
of a '(none) (current)'). This fixes this problem
and renames (none) to (provided by qube), because
that's what kernel = None should do.

fixes QubesOS/qubes-issues#7066