Warn about metadata leaks in backups
Add a message to warn the user that a backup always includes metadata
for all VMs on the system, to prevent e.g. leaking any sensitive qube
names, should the filtered down backup get shared with somebody.

The message is not shown when all VMs are selected, to make it more
noticeable when it could actually be relevant.

See also QubesOS/qubes-issues#2645