Fix cloning from non-default to default pool
When cloning a VM to the system's default pool, clone_vm() from
qubesadmin is invoked with pool=None. In that case, it is important to
set the pools (plural) argument to {} rather than the default value None
even though both are falsey, because clone_vm() takes the combination of
pool=None + pools=None to mean "use the source pools as the destination
pools".

The qvm-clone CLI already specifies pools={} correctly, so it was not
affected by the bug.

https://forum.qubes-os.org/t/templatevms-and-standalonevms-on-secondary-pool-cannot-connect-to-qrexec-agent/30811/15

Reported-by: https://forum.qubes-os.org/u/bengee
(cherry picked from commit f89632b0bdfa1739e4ba9fc646fc75c26c2c66be)